chore: refactor userauth
